fix: Aplicar tratativa de erros no merge entre a main e a branch feat/login
